packages/bot-cloud-handler/pyproject.toml
[tool.poetry]
name = "bot-cloud-handler"
version = "0.1.0"
description = ""
authors = ["Halk Liff <halkliff@pm.me>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://github.com/halkliff/EmaProject/tree/main/packages/bot-cloud-handler"
repository = "https://github.com/halkliff/EmaProject.git"
packages = [{ include = "bot_cloud_handler" }]

[tool.poetry.dependencies]
python = "^3.12"
Pillow = "10.2.0"
hachoir = "3.3.0"
cryptg = "0.4.0"
aiohttp = { extras = ["speedups"], version = "3.9.3" }
flake8 = "7.0.0"
black = "24.2.0"
ujson = "5.9.0"
python-dotenv = "1.0.1"
pyTelegramBotAPI = "4.16.1"
functions-framework = "3.5.0"
cloudevents = "1.10.1"
redis = "5.0.1"
pydantic = "2.6.2"
beanie = "1.25.0"
motor = { extras = ["srv"], version = "3.3.2" }
telethon = "1.34.0"
injector = "0.21.0"
icecream = "2.1.3"
types-ujson = "5.9.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.6.2"
commitizen = "3.16.0"
mypy = "^1.8.0"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["telebot.*", "telethon.*", "icecream.*"]
ignore_missing_imports = true

[tool.poetry.scripts]
export-requirements-files = "scripts.export_requirements_files:run"
start = "bot_cloud_handler:start"
# start dev command to run on your own command line to start the bot in watch mode,
# using node's nodemon
# start-dev = "npx nodemon --watch ./bot_cloud_handler --ext py --exec poetry run start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
